refactor: move all tags related functions
* Move all tags related functions into repository and service
* Call the repository from router as migration pattern
